tipc: copy SO_PRIORITY value to allocated skb's
This makes it possible to create mapping of priority to Ethernet 802.1p
(priority bits) using lartc or vconfig egress_map for TIPC user data traffic.

In addition to copying socket priority to the SKB in tipc_msg_build,
priority is also assigned to the allocated SKB's when:

-A new bundling buffer is allocated on the link
-Message fragmentation buffers are allocated
-Link changeover occurs and the packets buffered
 on the defunct link are duplicated to the new active link.
-Port level control messages are allocated

This change does not impact link level control messsages.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>