tipc: remove port lock
Still very much work in progress.
A number of issues need to be resolved:
Message rejection should be moved to link layer, and a
rejection queue added to the link.

Solve the problem with link_schedule_port, that
previously took the link lock from different contexts (both BH and user).

Most likely, there is more...

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>