tipc: add basic discovery mechanism
A simple translation mechanism from ipv4/ipv6 address
to TIPC port ID is introduced.
When a packet is received from a peer node, we calculate
a hash over the source IP and saves the ip and TIPC port ID
in a neighbor list. When packets are sent, we first try to
find an entry in the neighbor list and send the message directly
to the remote port. If no entry is found, the message is
sent out in a multicast manner.
ARP, NDP and DAD is disabled on the virtual device.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>