tipc: out-of-sequence packets must not be subject to GRO
The link layer does not know the sequence numbers of
each packet that a GRO chain is built up of, and must
assume that they are sequential.
Parallell streams over the same link will be interleaved
at the link-layer. If an interleaved packet is received
on the interface, we cannot treat subsequent packets from
the previous stream as the same flow.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>